+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
  * integrated puli/asset-plugin in puli/manager
  * added `ConfigManager` interface
  * added argument `$raw` to the getters in `ConfigManager`
+ * added `PuliEvents::PRE_BUILD_REPOSITORY` and `PuliEvents::POST_BUILD_REPOSITORY`
 
 * 1.0.0-beta4 (2015-04-13)